+++ src/util/MultiIndex.h
@@ -1,14 +1,15 @@
 #ifndef MULTIINDEX_H
 #define MULTIINDEX_H
 
-#include <cstddef>
 #include <array>
+#include <cstddef>
+#include <cstdint>
 
 namespace tndm {
 
-template<std::size_t D>
-int flatten(std::array<int,D> const& idx, std::array<int,D> const& shape) {
-    int flat = 0;
+template <std::size_t D>
+uint64_t flatten(std::array<uint64_t, D> const& idx, std::array<uint64_t, D> const& shape) {
+    uint64_t flat = 0;
     for (int d = D-1; d >= 0; --d) {
         flat *= shape[d];
         flat += idx[d];
@@ -16,9 +17,9 @@ int flatten(std::array<int,D> const& idx, std::array<int,D> const& shape) {
     return flat;
 }
 
-template<std::size_t D>
-std::array<int,D> unflatten(int idx, std::array<int,D> const& shape) {
-    std::array<int,D> a;
+template <std::size_t D>
+std::array<uint64_t, D> unflatten(uint64_t idx, std::array<uint64_t, D> const& shape) {
+    std::array<uint64_t, D> a;
     for (std::size_t d = 0; d < D; ++d) {
         a[d] = idx % shape[d];
         idx /= shape[d];
@@ -26,24 +27,25 @@ std::array<int,D> unflatten(int idx, std::array<int,D> const& shape) {
     return a;
 }
 
-template<std::size_t D>
-std::array<int,D> operator+(std::array<int,D> const& lhs, std::array<int,D> const& rhs) {
-    std::array<int,D> result;
+template <std::size_t D>
+std::array<uint64_t, D> operator+(std::array<uint64_t, D> const& lhs,
+                                  std::array<uint64_t, D> const& rhs) {
+    std::array<uint64_t, D> result;
     for (std::size_t d = 0; d < D; ++d) {
         result[d] = lhs[d] + rhs[d];
     }
     return result;
 }
 
-template<std::size_t D>
-std::array<int,D> operator-(std::array<int,D> const& lhs, std::array<int,D> const& rhs) {
-    std::array<int,D> result;
+template <std::size_t D>
+std::array<uint64_t, D> operator-(std::array<uint64_t, D> const& lhs,
+                                  std::array<uint64_t, D> const& rhs) {
+    std::array<uint64_t, D> result;
     for (std::size_t d = 0; d < D; ++d) {
         result[d] = lhs[d] - rhs[d];
     }
     return result;
 }
-
 }
 
 #endif // MULTIINDEX_H
